+++ Sources/UtiliKit/Instantiation/UIStoryboard+Extensions.swift
@@ -10,7 +10,7 @@ import UIKit
 public extension UIStoryboard {
     
     /// A small struct used to represent the name and bundle of a storyboad object.
-    public struct Identifier {
+    struct Identifier {
         public let name: String
         public let bundle: Bundle
         
@@ -40,7 +40,7 @@ public extension UIStoryboard {
     
      - Returns: A view controller of type T.
     */
-    public func instantiateInitialViewController<T: UIViewController>() -> T {
+    func instantiateInitialViewController<T: UIViewController>() -> T {
         guard let vc = instantiateInitialViewController() as? T else {
             fatalError("Unable to instantiate the initial view controller as \(T.self) from storyboard \(self)")
         }
@@ -54,7 +54,7 @@ public extension UIStoryboard {
      - Parameter element: An element of the appropriate type, used to configure the view controller.
      - Returns: A view controller of type T.
      */
-    public func instantiateInitialViewController<T: UIViewController & Configurable>(configuredWith element: T.ConfiguringType) -> T {
+    func instantiateInitialViewController<T: UIViewController & Configurable>(configuredWith element: T.ConfiguringType) -> T {
         let vc: T = instantiateInitialViewController()
         vc.configure(with: element)
         
@@ -66,7 +66,7 @@ public extension UIStoryboard {
     
      - Returns: A view controller of type T.
     */
-    public func instantiateViewController<T: UIViewController>() -> T {
+    func instantiateViewController<T: UIViewController>() -> T {
         guard let vc = instantiateViewController(withIdentifier: T.storyboardIdentifier) as? T else {
             fatalError("Unable to instantiate a view controller with identifier \(T.storyboardIdentifier) from storyboard \(self)")
         }
@@ -80,7 +80,7 @@ public extension UIStoryboard {
      - Parameter element: An element of the appropriate type, used to configure the view controller.
      - Returns: A view controller of type T.
      */
-    public func instantiateViewController<T: UIViewController & Configurable>(configuredWith element: T.ConfiguringType) -> T {
+    func instantiateViewController<T: UIViewController & Configurable>(configuredWith element: T.ConfiguringType) -> T {
         let vc: T = instantiateViewController()
         vc.configure(with: element)