+++ Sources/UtiliKit/Instantiation/UITableView+Extensions.swift
@@ -15,7 +15,7 @@ public extension UITableView {
     
      - Parameter type: The type of the cell being registered.
     */
-    public func register<T: UITableViewCell>(type: T.Type) {
+    func register<T: UITableViewCell>(type: T.Type) {
         register(type, forCellReuseIdentifier: T.reuseIdentifier)
     }
     
@@ -24,7 +24,7 @@ public extension UITableView {
     
      - Parameter type: The class of the nib being registered. This should match its reuse identifier.
     */
-    public func registerNib<T: UITableViewCell>(for type: T.Type) {
+    func registerNib<T: UITableViewCell>(for type: T.Type) {
         register(UINib(nibName: T.nibName, bundle: Bundle(for: type)), forCellReuseIdentifier: T.reuseIdentifier)
     }
     
@@ -33,7 +33,7 @@ public extension UITableView {
     
      - Parameter type: The type of the class being registered.
     */
-    public func registerHeaderFooter<T: UITableViewHeaderFooterView>(type: T.Type) {
+    func registerHeaderFooter<T: UITableViewHeaderFooterView>(type: T.Type) {
         register(type, forHeaderFooterViewReuseIdentifier: T.reuseIdentifier)
     }
     
@@ -42,7 +42,7 @@ public extension UITableView {
     
      - Parameter type: The class of the nib being registered. This should match its reuse identifier.
     */
-    public func registerNib<T: UITableViewHeaderFooterView>(forHeaderFooterView type: T.Type) {
+    func registerNib<T: UITableViewHeaderFooterView>(forHeaderFooterView type: T.Type) {
         register(UINib(nibName: T.nibName, bundle: Bundle(for: type)), forHeaderFooterViewReuseIdentifier: T.reuseIdentifier)
     }
     
@@ -53,10 +53,24 @@ public extension UITableView {
      - Parameter indexPath: The index path of the cell.
      - Returns: A table view cell of type T.
     */
-    public func dequeueReusableCell<T: UITableViewCell>(for indexPath: IndexPath) -> T  {
+    func dequeueReusableCell<T: UITableViewCell>(for indexPath: IndexPath) -> T  {
         guard let reusableCell = dequeueReusableCell(withIdentifier: T.reuseIdentifier, for: indexPath) as? T else {
             fatalError("Could not dequeue a reusable cell of type \(T.self) with identifier \(T.reuseIdentifier) for use in \(self)")
         }
+        
+        return reusableCell
+    }
+    
+    /**
+     Returns a reusable table view cell of type T.
+     
+     - Parameter indexPath: The index path of the cell.
+     - Parameter element: An object used to configure the newly dequeued cell.
+     - Returns: A table view cell of type T configured with `element`.
+     */
+    func dequeueReusableCell<T: UITableViewCell & Configurable>(for indexPath: IndexPath, configuredWith element: T.ConfiguringType) -> T  {
+        let reusableCell: T = dequeueReusableCell(for: indexPath)
+        reusableCell.configure(with: element)
         
         return reusableCell
     }
@@ -66,10 +80,23 @@ public extension UITableView {
     
      - Returns: A headerFooter view of type T.
     */
-    public func dequeueReusableHeaderFooterView<T: UITableViewHeaderFooterView>() -> T {
+    func dequeueReusableHeaderFooterView<T: UITableViewHeaderFooterView>() -> T {
         guard let reusableView = dequeueReusableHeaderFooterView(withIdentifier: T.reuseIdentifier) as? T else {
             fatalError("Could not dequeue a reusable view of type \(T.self) with identifier \(T.reuseIdentifier) for use in \(self)")
         }
+        
+        return reusableView
+    }
+    
+    /**
+     Returns a reusable headerFooter view of type T.
+     
+     - Parameter element: An object used to configure the newly dequeued header or footer view.
+     - Returns: A headerFooter view of type T configured with `element`.
+     */
+    func dequeueReusableHeaderFooterView<T: UITableViewHeaderFooterView & Configurable>(configuredWith element: T.ConfiguringType) -> T {
+        let reusableView: T = dequeueReusableHeaderFooterView()
+        reusableView.configure(with: element)
         
         return reusableView
     }